requirements.txt
# AI Trading Bot - Dependencies
# Version 1.0

# Core
python-dateutil>=2.8.2
pyyaml>=6.0
python-dotenv>=1.0.0

# Data & Numeric
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Technical Analysis
ta>=0.10.2
pandas-ta>=0.3.14b

# Machine Learning
scikit-learn>=1.3.0
lightgbm>=4.0.0
catboost>=1.2
xgboost>=2.0.0

# Deep Learning (optional, for LSTM/Transformer)
torch>=2.0.0

# Exchange API
ccxt>=4.0.0
websockets>=11.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0
alembic>=1.11.0

# Async
aiohttp>=3.8.0
asyncio-throttle>=1.0.2

# Monitoring & Logging
loguru>=0.7.0
prometheus-client>=0.17.0

# Dashboard
dash>=2.11.0
plotly>=5.15.0

# Notifications
python-telegram-bot>=20.0
aiosmtplib>=2.0.0

# Testing
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
isort>=5.12.0

# Backtesting
vectorbt>=0.25.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
cryptography>=41.0.0
